packaging: handle new NFS installation source with inst.stage2=nfs:...
This block of code runs when you're setting up a new NFS installation
source (i.e. method.method is 'nfs'). At this point 'device' is the
mount device for DRACUT_REPODIR.

If you have "inst.stage2=nfs:host:/path", 'device' is going to be
"host:/path", and iutil.parseNfsUrl('nfs:%s' % device) yields
host="host" and path="/path".

(If you have "inst.stage2=hd:sdX:..." you'll get host="sdX" and
path="path"; otherwise you probably won't have DRACUT_REPODIR mounted
and thus no 'device'.)

The problem, then, is this: if you have 'path', then 'path in device' is
always true, because 'path' was just split out of 'device'. Which means
we'll always assume that DRACUT_REPODIR is the mount we want, even if
it's not!

To fix this, we need to actually check to see if the thing mounted at
DRACUT_REPODIR is actually the same as the thing in our 'method' object.
If we have the same host and path, we can just use the mount as-is,
otherwise we should mount the newly-provided source.

(forward-ported from rhel7-branch commit 58402290)