Changes for scheduling size change on an existing device (#1076055)
Related: rhbz#1076055
Related: rhbz#1196721

(cherry-picked from commit cb4794835fbf97cd51275f61f82ec428db42f1a4)

When changing size, the device which gets its size changed is always the
raw device.

Do a little ad-hoc adjustment of the size in case the raw device and the
device that the user sees the size of are different.
Since the only situation where these are different is with a LUKS device
use crypto constant for the adjustment.

Signed-off-by: mulhern <amulhern@redhat.com>