Refactor handling of fsCombo considerations.
Previously there was a method called _resolve_btrfs_restrictions().
Now there is a method called _update_fstype_combo() which behaves
slightly differently.

Changes are:

* Make the method dependent on the device_type chosen, rather than
a boolean result. Everything about the fsCombo changes is really
just dependent on this value, so it's better to make that explicit by the
parameter.
* Remove should_be_btrfs and make whether or not to display and lock
btrfs in fsCombo dependent only on whether btrfs device type is selected.
Checking whether btrfs is supported in this method is a bad idea, because:
 - If the device already exists and is btrfs, then the fsCombo should
be locked to btrfs and nothing else.
 - If the device type has already been changed to btrfs, then only btrfs
filesystem is allowed, so locking the fstype to whatever the user had
previously chosen before they made the device type change isn't really
much better than displaying btrfs filesystem and locking it to that.
Neither one is something that anaconda/blivet can ultimately do.
* Move setting of sensitivity of fsCombo in with the other manipulations
of fsCombo. They are all connected.
* Make method slightly more robust. There are a bunch of situations that,
by close reasoning, you can prove could not happen in the current GUI,
but that could change. So, situation where device type is now BTRFS and
btrfs entry is already in fsCombo is handled. Situation where device type is
not BTRFS and btrfs is not in fsCombo is handled explicitly, rather than
implicitly. Situation where btrfs is in combo and some other fs that is
past btrfs in the combo is selected is now handled, so that user won't see
selected fs spontaneously shift to a different one on btrfs removal.

Signed-off-by: mulhern <amulhern@redhat.com>