Add transifex branch check to makebumpver
When makebumpver runs it will check the .tx/config file for a section
named anaconda.<branch> and for one named main. If it doesn't find the
expected anaconda branch or it finds any extras it will exit with a 1.

This can be overridden by passing --skip-tx or using make
scratch-bumpver.

The goal here is to catch using the wrong transifex branch as soon as
possible since it can result in lost translations like rhbz#1138644