Allow kickstart with no method (#972265)
If the kickstart has no method allow Anaconda to fallback to the closest
mirror or DVD.

Also change the source spoke to be complete for an automated install if
it has a baserepo, not a baserepo and a method. If it has a baserepo
then that means that the method setup, whatever it is, is ok.