Omit calculation and use of active_dev_type.
It is only calculated for logging.

itr is only used to calculate active_dev_type, so get rid of that
calculation as well.

Note that the calculation of active_dev_type is mostly a copy of the body
of _get_current_device_type(), except that it uses combo, instead of
self._typeCombo. But, for the method to make any sense, combo must
be self._typeCombo, so put that in the preconditions.

Since it is nicer to print readable names than integer codes, though,
use get_current_device_type_name() to get that name and use in log.debug
statement.

Then transform it to integer code for blivet method calls.

Signed-off-by: mulhern <amulhern@redhat.com>