Rewrite _bound_size() to bound_size() in storage_utils.py
Fix a bug where the possibility of maxSize() being 0, a value which is used
to mean no maxSize() given, is not taken into account.

Since, a fall back is needed in case this happens, pass old_size to
_bound_size as well.

Rearrange if statement a bit to make it clearer now that additional
possibilities are taken into account.

Do a bit more logging, in case something useful-to-know or bad happens.

Signed-off-by: mulhern <amulhern@redhat.com>