rpmostreepayload: Copy all subdirectories of /usr/lib/ostree-boot
Previously we were special casing extlinux/ and grub2/, but this
misses out on efi/.  To support rpmostreepayload + GRUB2 + EFI, we
need special handling here.  Any bootloaders installing files in /boot
instead of a subdirectory will then need to be special cased (or
"fixed").

Now in a perfect world, I'd have a time machine and we could go back
and have every bootloader install in a subdirectory /boot/data or
something.  But we basically can't change existing systems, which kind
of traps us into the current RPM-level layout forever, with rpm-ostree
postprocessing it.

(One could argue it's rpm-ostree's job to postprocess this so Anaconda
 wouldn't have to care.  Open to that.)

Related: rhbz#1153409