rpmostreepayload: Rework remote add handling
**NOTE** This will break rpmostreepayload unless you have
a version of ostree later than 2014.12 with
https://git.gnome.org/browse/ostree/commit/?id=f6a6e68412c9f1be2b5de1ded79b92aa340ab22c

OSTree's handling of remotes is a bit funky, because it supports
non-root operation, where the remotes are stored in the repo config,
and historically that was the only supported method.

But more recently, support for /etc/ostree/remotes.d was added; it's a
lot friendlier to system administrators for the special case of the
system repository /ostree/repo.

However for Anaconda, there's a bootstrapping problem as that
directory doesn't yet exist when we're doing an install.  And we
(fortunately) don't even try to write there because OSTree doesn't
understand that /mnt/sysimage/ostree/repo is a system repository.

The goal is for the remote to end up in /etc/ostree/remotes.d, so to
do that unfortunately involves the complex dance in this patch.