bootloader: Bridge efi_dir configuration earlier for rpmostreepayload
rpmostreepayload wants a stub bootloader configuration to be written,
then it updates it.  There is a bit of magic in the middle of
bootloader.py to bridge the install class' efi_dir to the bootloader
configuration - we just need to set this up before calling
writeBootLoaderFinal().

Otherwise the config ends up in BOOT, which doesn't work.

Related: rhbz#1153409