Do not install interactive exception handler in cmdline mode (#1155979)
The cmdline mode is, by definition, non-interactive so it doesn't make much
sense to use an interactive exception handler for it. However, we need to debug
the issues with cmdline mode sometimes so it is useful to have a way to force
the interactive exception handler even for the cmdline mode.