Make anaconda more scrollable (#1135024)
Add scrolled windows to spokes that don't fit so good on screens with a
low vertical resolution. This change adds a scrolled window to the now
taller left side of the custom partitioning screen and around the entire
source selection screen. The 24px right padding on the GtkAlignment is
moved to a right margin on the viewport so that the scroll bars appear
all the way to the right.

(cherry picked from commit 2617ed8fa3209cae69225eb2a4f26a57913f5513)