Look up most widgets relative to the currently displayed screen.
With the move to a GtkStack, spokes and hubs get displayed and then when
the user leaves, they live on.  This means that widgets like the Done button
exist in multiple places as far as dogtail knows, but only one of them is
actually valid.  The rest are not displayed, thus have invalid coordinates,
thus will fail when a click is attempted.