Rework the placement of items on hubs.
We have been using xscale=0.5 on the hub alignment to center and pad
items when the alignment's children do not use the entire width of the
hub. This has a couple of problems. When the alignment does not need a
lot of space, the scrollbar, if shown, is way over to the left instead
of all the way on the right side of the screen. If the alignment does
need a lot of space, there's no padding so the spokes are pushed all the
way against the left sidebar and the right edge of the window.

Redo the alignment of the hub GtkAlignment to match the one on the
spokes: xalign=0.5, xscale=1.0. Add left and right padding to the
alignment to keep things away from the sidebar and the edge of the
window.