Allow adding prepboot to a blank disk in custom (#1155660)
On systems where the stage1 device can only be a partition (PPC for
example) the self.data.bootloader.bootDrive isn't setup until you add a
prepboot partition. If you start with a totally blank disk and added a
prepboot partition first the accordion wouldn't be populated, resulting
in a traceback.

This changes custom to allow biosboot and prepboot to be added when
self.data.bootloader.bootDrive hasn't been setup yet.

This would also happen on UEFI when adding biosboot or prepboot as the
first partition to a blank disk.