custom: Clearing errors should also clear Done clicked state (#1158609)
When there is an error we tell the user to click done again to exit. If
they do anything else the state needs to be cleared. We already call
clear_errors from all these places in order to reset the error bar, so
add clearing of the done clicked state.

Since do_check also clears errors, set the state to True *after* an
error has been found.

Resolves: rhbz#1158609