Provide useful hints on TTY1 during the installation
Until now TTY1 was not really used for anything other than printing
current Anaconda version. This has been changed and the Anaconda
version is now followed by useful hints for Anaconda users with
special emphasis of Anaconda debugging (where are the log files stored,
how can I get a shell, etc.).

The hints are also parametrized to be displayed only when appropriate
(no need to tell users that Anaconda has a text mode if they already
are in textmode, etc.).