Revert "Revert productName repo name change (#1128474)"
This reverts commit 51f204b9694400b512d0e8cf33c4b409ac0fa0d4.

Yes, this is absurd. But this will be the last time I change this. The
rule will now be:

Include a repo in your installation environment that has an id
matching the lower case first part of your product name when split at
the '-' character.

If you don't do that closest mirror will not work.

For Fedora products this means including ONE repo with an id of
'fedora'.