Allow specifying which of the GUI tests you want to run.
When there was only one, it was easy.  However for developing new tests,
it's helpful to just run the new one by itself.  This allows doing so by
setting the GUI_TESTS environment variable.

While I'm at it, add a couple messages to explain why the whole test suite
did not get run.