Add a basic test of the reclaim dialog.
This only tests the situation where you've got a disk with one partition,
with a filesystem on that taking up the entire disk.  That filesystem can
be deleted and then anaconda will have enough space to keep going.

There are certainly more tests that could be done with this dialog, and
that's what needs to be worked on next after verifying running multiple
tests at the same time works.