Disable payloads that failed to setup (#1162732)
Add a unsetup method to Payload that is called by PayloadManager when
setting up a payload fails. For dnf and tarpayload, just unset
some attributes during unsetup, for live unmount the tree, and for
yumpayload unset the _setup flag and expand the flag's use to disable
access to the yum comps magic properties.

This fixes a problem that is most troublesome with yum, where the setup
for a kickstart or command-line repo can fail, and a later variable
access in the software spoke will start yum downloading metadata from
somewhere else and setting half of the variables that we expect and
crashing. By invalidating the payload we avoid the accidental fallback
to the wrong repos.

(cherry picked from commit 27e9d4af2cf3e0b832ac7c70b243b53f7c034641)