Support high contrast mode in fedora-welcome (#1160499)
Thanks to Matthias Clasen for fixing the icon. Remove the hardcoded text
color from the icon labels and use a named icon for "Install to Hard
Drive" instead of a full path.

Solve the problem of themeing the fedora-welcome launcher icon by
removing the launcher icon. This program is not user-launchable so
there's no reason to display a fedora logo for it.

Switch to the branded anaconda logo for the liveinst desktop icon.
Remove the generic logo.