Related bug can have different fixed-in and state
If a related bug is the same as the bug listed in the summary it
overrides the need for a matching fixed-in or state. eg. a fix to a
patch previously pushed that is already ON_QA or has an older fixed-in
version.

If the bug in the summary is different from the related bug that implies
Resolved and it must have correct state, fixed-in and acks.