Add support for custom gid to advanced user setup (#1163803)
The dialog for the advanced user setup claims you can specifiy the group
id like this: groupname (1234), but support for that was never actually
added.

This adds support for parsing the gid, as well as tests for the new
parse function. Note that if it cannot parse the group entry it will
ignore it. At some future time we should rework error handling in this
dialog to make this more user friendly.

Resolves: rhbz#1163803