Don't use blivet in the gui tests.
First off, I think it's definitely more powerful than is going to be needed
for these tests.

Second, I think the model of having a comprehensive view of the storage on the
system combined with running multiple tests at once is causing lots of problems
and has the potential to cause a lot more.

Instead I'm just going to make filesystems using system tools which has the
advantage of not leaving anything around for another test to pick up.