datetime_spoke: Fix warnings about removing nonexistent source
First, we missed setting the variable to None in one place for the
update source.

Second, when the "start updating" source returns False, we need to set
the corresponding variable to None as well in the source itself to
preserve the invariant that the ID is non-None if and only if the
source is outstanding.