Handle unstaged changes in the pre-commit hook.
Instead of failing pre-commit when there are unstaged changes, create a
temporary file containing only the staged changes and run pylint on
that. pylint may fail if there are both staged and unstaged changes in
multiple that depend on one another, but this will at least make it more
likely that commit --interactive will work.