Add help content subpackage (#1072033)
Add the anaconda-user-help subpackage to host the content
for the built-in help system.

The subpackage is built from a separate tarball that contains the
RHEL7 installation guide and a build script that generates the
Anaconda help content from the installation guide files.

The sub package also has its own version, as it should be eventually
split to a separate package, so that there are no weird version number
jumps once that happens.

The version number is based RHEL version being targeted by the help
content. For example anaconda-user-help-7.1.1 means that it is
targeting RHEL7.1 and that it is the first release of the help content
package for this RHEL version.

Related: rhbz#1072033