rpmostreepayload: Avoid shutil.copytree in favor of cp -r to fix symlinks
shutil.copytree() follows symlinks by default which I didn't realize,
and this actively breaks things as grubenv is a symlink.

Since we're shelling out to 'cp' in a case above, might as well do it
here.

This also means that if we hit e.g. ENOSPC we'll enter the error
handling codepath instead of an uncaught traceback.

(Yes, this code is crappy, should probably live in bootloader.py,
 except I'm going through gyrations here to avoid requiring
 any changes in how mainline bootloaders work)