Shrink the LV if the VG is out of space (#1153376)
lvm doesn't always create the sizes that we expect it to. This can lead
to errors like this:

Volume group "VolGroup" has insufficient free space (1965 extents): 1968
required.

This change checks the LV's size against the VG's actual free space and
shrinks the LV to fit if it would be to big.

This is a port of commit 0c40779fa8f from python-blivet.

Resolves: rhbz#1153376