Fix EOF error that occurs if user input required in x3270. (#1171135)
Prior to fixing rhbz#1040933, anaconda was always launched from a tmux
session. Now however, if you install via text/cmdline kickstart, sshd
is suspended and everything is dumped to the x3270 console.

If it happened that user input was required (e.g. if you supplied an
incomplete ks file), this EOF error was triggered when raw_input() was
called from pyanaconda.ui.tui.simpleline.base, because there was no std
input device for it to open. Well, standard input on x3270 is a tty, so
define that here to prevent the traceback and allow users to provide
input if it is needed.

Resolves: rhbz#1171135

Signed-off-by: Samantha N. Bueno <sbueno+anaconda@redhat.com>