Simplify and robustify handling of fstype combo box.
Changes are:
* just fill the combo from scratch rather than traversing it to remove
unwanted items.
  - therefore, don't fill the combo in _initialize
* Handle possible BTRFS device type in _setup_fstype_combo().
* Sort _all_ the names alphabetically, instead of tacking on a few
  unsupported filesystem names at end.
* Make _fs_types attribute be a set, not a list.
* Make a note that _update_fstype_combo() is idempotent and needs to stay
that way.

Non-changes are:
* Retain the complicated check for the original format's name. I don't
know why the check is all about the format's type instead of the name,
but it is, so I figure I'ld better leave it like that.

Advantages are:
* Get rid of a code smell. The first half of the code that removed
unsupported filesystem names assumed unsupported
filesystems names could occur anywhere in the combo box,
while the second part of the removal code assumed that unsupported
filesystem names could only occur after the device's filesystem name.
In the current setup, the second assumption is the correct one, so the
code actually worked. If the first assumption had been correct, then
the second half of the code would have been buggy. Therefore, the whole
code block looked suspicious.
* Make the custom spoke a bit less stateful.
* Make sure that btrfs name is included correctly. It currently is,
but only because _setup_device_type_combo() is invoked after
_setup_fstype_combo(), and _setup_device_type_combo() causes
_typeCombo._setActive() to be invoked, which triggers an event which
causes on_device_type_changed() to be invoked, which in turn invokes
the _update_fstype_combo() method. It seems wrong
to expect memory of that ordering requirement to be retained forever,
so better to make it unnecessary by invoking _update_fstype_combo()
explicitly in _setup_fstype_combo(). This means that _update_fstype_combo()
is called twice, but since it is idempotent, that is no problem.
* Fewer lines of code.

Signed-off-by: mulhern <amulhern@redhat.com>