Fix issues with the date&time not being updated on timezone changes
We need to wait for the result of _set_timezone because we are interested in its
return value. And we should use timezone map to refresh the selected timezone as
its callbacks update the rest of the spoke correctly.

(port of c7442823e88aa3a164dddf645032e2bcacdf1df0 from rhel7-branch)