Fix issues with the date&time not being updated on timezone changes
We need to wait for the result of _set_timezone because we are interested in its
return value. And we should use timezone map to refresh the selected timezone as
its callbacks update the rest of the spoke correctly.

Resolves: rhbz#1267450

(cherry picked from commit 5173d97a76d62f13966897875a3028c2e76ba0fd)