Fix threading issues for dasdfmt in gui storage. (#1073982)
Prior code I committed manually issued a call to spoke.execute()
in the storage spoke once dasdfmt had been run (if doing a ks
installation). Turns out this is a Bad Idea. spoke.execute() is
called anyway from pyanaconda.ui.gui.hubs once it is detected that
a spoke has changed, so this is not necessary.

Resolves: rhbz#1073982