Reorganize run_kickstart_tests.sh a bit to use parallel.
Turns out it wasn't as hard as I'd thought it would be.

Also note that for now I am still restricting the number of jobs running
in parallel to two, because each needs a lot of memory and I don't want to
assume too much about the test host.  I am open to ideas about better ways
to do this.