Ellipsize comboboxes (#1170275)
This way super-long container types (LVM Thin Provisioning, translations
of LVM Thin Provisioning), RAID descriptions and container names won't
blow out the window size. For container names, ellipsize in the middle
instead of the end so that supercrazylongcontainername-data01 is easily
distinguishable from supercrazylongcontainername-data02. In all cases
the full string is visible when the ComboBox is expanded.