Start the network before the display (#1167103)
Sometimes the display is on the network.

If there is a kickstart file containing a static network configuration,
and dracut started the network using dhcp (for example, vnc was on the
command line but no ip arguments were), it's possible to change
addresses after VNC has started, which doesn't work real well.