Remove allowDirty parameter from mountExistingSystem() call.
mountExistingSystem() has been re-written so that it behaves as if
allowDirty is always True and the parameter has been removed.

Also, remove try/except/else block for DirtyFSError. DirtyFSError is only
thrown when allowDirty is False.

This looks like a lot of changes, but everything below the except block is
just one big outdent of the body of the else block.

Update blivet requires for change in mountExistingSystem().

Signed-off-by: mulhern <amulhern@redhat.com>