Updates for new Size.convertTo() spec.
The occurrance of convertTo() in custom.py is really unnecessary, so it
is removed. The comment just above is updated to be consistent with the
code below. The if-condition ends up being overall a bit different, but the
whole if-statement has just the same effect as before.

Also, according to qemu-img documentation, sizes are in power of 2 units,
not SI units, so change in tests.

Signed-off-by: mulhern <amulhern@redhat.com>