Move environment group selection logic to PackagePayload (#1179362)
The selectEnvironment method in yum only iterates over the groups within
an environment and selects each, which is possible using the existing
PackagePayload API. selectEnvironment in dnf selects the environment
using dnf.Base.environment_install, which does not change the kickstart
data and is probably premature. Replace both methods with one that calls
selectGroup on each group in the environment.