Animate the screen transitions.
This emphasizes the relationship between hubs and spokes. When entering
a spoke, the screen slides up to reveal the spoke being entered. When
returning to a hub, the hub slides down over the spoke being finished.
For hub->hub (or other standalone) transitions, the old screen slides
left (or right, in RTL languages) to reveal the new screen.

Thanks to Ryan Lerch for thinking of the idea and Máirín Duffy for
coming up with transition types that look nice.