Use threadMgr to wait for exception handling to finish
If an exception appears in a non-main thread prior to GUI being started,
python-meh or our own error handling code start a Gtk main loop in such a
thread. We need to wait for such main loop to finish instead of running another
one in the main thread. But even if python-meh or our error handling code call
sys.exit(), the process doesn't exit because the main thread is still running
and it is not a daemon thread. Thus we need to wait for the exception handling
code to finish and then exit the process from the main thread.