Maximize anaconda instead of running fullscreen (#1164457)
A fullscreen anaconda is inconvenient when running outside of the
regular installer media, since on Live installs some things are only
configurable in the desktop environment, and on dir and image installs
you might have other things to do while anaconda is running. Instead,
maximize anaconda and use the hide-titlebar-when-maximized property to
hide the titlebar on window managers that support it, which is really
just GNOME.

Add an option to GraphicalUserInterface to start with a fullscreen
anaconda for downstream users of the GUI that would prefer that
behavior.