Run AnacondaExceptionHandler in cmdline mode (#1256385)
cmdline mode still needs to have the exception handler run. It exits
anaconda when there are problems, without it you can get stuck in a loop
in the summary hub when there are incomplete spokes.

This adds an interactive argument to it so that when the handler runs in
cmdline mode it only prints out the error and calls sys.exit(1) --
skipping the call to the superclass from meh which tries to run a UI to
prompt for debugging.

(cherry picked from commit 98ec24be63dcc66de695dbfc86ef0fb08ab31f01)

Resolves: rhbz#1256385