Hook up jenkins support into makebumpver.
First, add a line to your ~/.rhbzauth file that points to the json API on your
jenkins server:

    JENKINS=http://jenkins.wherever.com/api/json

If this is not set, then the jenkins check will be skipped.

Then, make sure the name being passed to makebumpver is the same as the
beginning of some jenkins job you want to check, and that self.git_branch
is the end of the same jenkins job.  Naming is important.

For now this is only in an advisory role - jenkins will be checked and if the
tests do not pass, a warning will be logged.  But it will not stop the build.