Fix an issue in the previous pre-existing repo kickstart patch.
The previous patch was intended to allow you to do this:

    repo --name=updates

And then anaconda would read the pre-existing updates repo out of /etc
and enable the repo.  This worked, but it had one fatal flaw:  because we
read all the repos off disk, you couldn't then add your own repo with the
same name as one of those pre-existing ones.  This isn't normally a problem
except for livecd kickstarts, and they do this all over the place.

In yumpayload.py we solve this by letting a new repo delete the definition
it would conflict with, so we need to do that in dnfpayload.py too.