Restrict payload kernel versions to kernels in the payload (#1074358)
Remove the generic implementation of kernelVersionList from Payload.
Instead, have each payload return an iterable of the kernels that were
actually in the payload. Add a method to the packaging module to handle
version comparisons in the non-package payload types.

Remove rescueKernelList, since the bootloader was the only thing to use
this, so the bootloader can look for resuce kernels on its own.