Don't process continue-clicked events for windows that aren't shown.
This is another tactic in trying to fix the frequent kickstart problem of
processing a continue-clicked event twice:  once when the summary hub is up,
and then very shortly after when the progress hub is up.  This fix will
ignore any events where the window we should be operating on is not the one
that is currently displayed.