Add a method to set environment variables for child processes
iutil.setenv will manipulate a dictionary that is added to a copy of
os.environ that is passed to Popen. This way environment variables can
be made available to child processes without modifying the environment
of the anaconda process, which is not thread safe.

Use iutil.setenv in place of os.environment for a variety of environment
variables not needed by the anaconda process.