Make it possible to ignore individual newly added dependencies
The -f flag makes it possible to automatically fetch packages
for newly added dependencies. This usually works fine as long as the
packages are available from Koji. But in case the given dependency is
not yet available, the slow Koji package lookup holds back every such
makeupdates run considerably.

So add a new -i/--ignore option that makes it possible to ignore individual
package dependencies. Both exact match and globs are supported.