Avoid setting $LANG and $LANGUAGE, except where we can't
Remove the language argument from AnacondaBaseWindow.retranslate.
Setting the locale and language environment is handled by the caller, so
retranslate should only refresh translations to the current locale.

Setting $LANG is needed by Python's gettext module, so we can't get rid
of all of it. Handle the UEFI firmware language by doing it earlier, and
handle GUI changes by setting the envrionment variable in a way that
kinda maybe sort of looks thread-safe.