Add a pylint module to look for modifications to the environment
The environment cannot be safely modified in a multi-threaded program,
because it has the potential to crash getenv calls, and we can't control
calls to getenv because basically everything uses getenv.

For the easy fixes, mark setupEnvironment as safe since it's run before
threads are started, mark the environment cleanup in anaconda-yum as
safe since it's run in its own, single-threaded process, and remove the
environment cleanup at the end of handle-sshpw because it doesn't do
anything.