Set $LIBUSER_CONF early
The contents of the config file may change after the payload is setup
and packages are installed, but the environment variable needs to be set
early. For that reason, treat any failure to access an existing
libuser.conf as an error in createLuserConf.

(cherry picked from commit 637682b6c15dbe2b73e7e1f52a4f8efece7410b7)

Related: rhbz#1188287