Clear the default titlebar text
By default the Anaconda/Initial Setup window titlebar (when it is visible)
displays __init__.py in the titlebar (name of the file where the GTK Window
class is instantiated). This looks ugly and is not really needed as we
always have a in-window title for all windows, so just set the titlebar text
to an empty string.