Support kickstart test jobs out to multiple computers with parallel.
Basically, you pass TEST_REMOTES= as a space separated list of hostnames to
run on.  You will need to have root access on each of those hosts, and root's
ssh keys will need to be set up to allow logging in without a password.  lmc
requires running as root.

Do not add localhost to TEST_REMOTES.  It will be added automatically.

Adding this support required breaking run_kickstart_tests.sh out into a script
that does the work of calling lmc, and a script that interfaces with make and
fires off parallel.  This is because parallel really wants a single command to
run, and having the part that does the actual work be bash function in
run_kickstart_tests.sh was not working out.

A later commit is going to add an awful lot of documentation here.