Fix package and group removing with the dnf payload.
We can't really use dnf the same way we were using yum here.  With yum, we were
treating the transaction as a thing we could add and remove stuff from as much
as we wanted and then when we were done, tell yum to go ahead and do it.  dnf
doesn't work that way - we need to keep track of the additions and removals in
anaconda, and then when we are done tell dnf the result.

Thus:

(1) For groups, we implement selectEnvironment ourselves.  Just check that the
group is not in the excluded list when adding.  Same for when iterating over
the list of groups to add.

(2) For packages, make sure they're not in the excluded list when adding.  For
a package to remove that is part of a group we are installing, we are already
passing the excluded list along to dnf.  That should work.

(3) Remove _remove_package and _remove_group, since they don't (and won't)
work.