Make print statement print output w/out surrounding parentheses.
Previously print statement was:

print dev, size
yielding outtput like: sda 8394.0

it was changed to
print (dev, size)
which was interpreted by Python 2 as a print statement on a tuple, and
displayed accordingly,
('sda', 8394.0)
while Python3 interpreted as a print call on two arguments and displayed
sda 8394.0.

Use a format string, so that Python 3 and Python 2 will display consistantly
with each other and with previous version.

The alternative is to import the print function from the future, but that
seems like more trouble.

Signed-off-by: mulhern <amulhern@redhat.com>