Handle NoneType Exception after deleting PPC partitions (#1021445)
Users reported that the installer would crash when deleting all
existing partitions on the PowerPC Platform. This crash occurred
because the installer preferentially looks for a PRePBoot on the
same partition as /boot and, if it couldn't find any partitions
assigned to the /boot mount point, the object containing
boot devices would be null and the subsequent calls to that
object would trigger the failure. Modified the code such that
it checks if the value is NoneType and, if it is, simply skip
the codeblock that performs the preferential lookup. Updated the
type check for mdarray to validate the presence of the type
attribute so that, if the boot device object does not include
the type attribute, it will pick the sane default as intended.

Resolves rhbz#1021445