Add kickstart tests for an NFS instrepo and addon repos.
These tests requires an NFS-accessible mirror of the Fedora tree as well
as two specially constructed repositories to test the behavior of the
repo command. The contents of the repositories are described in the
comments in the kickstart files. Four new environment variables are used
to configure the servers and paths.